+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDK.java
@@ -74,6 +74,9 @@ public class SDK {
      * Endpoints for testing global parameters.
      */
     public Globals globals;
+    public Nested nested;
+    public NestedFirst nestedFirst;
+    public NestedSecond nestedSecond;
     /**
      * Endpoints for testing the pagination extension
      */
@@ -337,6 +340,12 @@ public class SDK {
 		this.generation = new Generation(this.sdkConfiguration);
 		
 		this.globals = new Globals(this.sdkConfiguration);
+		
+		this.nested = new Nested(this.sdkConfiguration);
+		
+		this.nestedFirst = new NestedFirst(this.sdkConfiguration);
+		
+		this.nestedSecond = new NestedSecond(this.sdkConfiguration);
 		
 		this.pagination = new Pagination(this.sdkConfiguration);