+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDKConfiguration.java
@@ -35,8 +35,8 @@ class SDKConfiguration {
 	}};
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
-	public String sdkVersion = "1.31.0";
-	public String genVersion = "2.91.2";
+	public String sdkVersion = "1.31.1";
+	public String genVersion = "2.91.4";
 	public Map<String, Map<String, Map<String, Object>>> globals = new HashMap<String, Map<String, Map<String, Object>>>(){{
 		put("parameters", new HashMap<String, Map<String, Object>>());
 	}};