+++ php-client-sdk/src/SDK.php
@@ -78,6 +78,12 @@ class SDK
      */
 	public Globals $globals;
 	
+	public Nested $nested;
+	
+	public NestedFirst $nestedFirst;
+	
+	public NestedSecond $nestedSecond;
+	
     /**
      * Endpoints for testing the pagination extension
      * 
@@ -170,6 +176,12 @@ class SDK
 		$this->generation = new Generation($this->sdkConfiguration);
 		
 		$this->globals = new Globals($this->sdkConfiguration);
+		
+		$this->nested = new Nested($this->sdkConfiguration);
+		
+		$this->nestedFirst = new NestedFirst($this->sdkConfiguration);
+		
+		$this->nestedSecond = new NestedSecond($this->sdkConfiguration);
 		
 		$this->pagination = new Pagination($this->sdkConfiguration);