+++ python-client-sdk/src/sdk/sdk.py
@@ -8,6 +8,9 @@ from .first import First
 from .flattening import Flattening
 from .generation import Generation
 from .globals import Globals
+from .nested import Nested
+from .nested_first import NestedFirst
+from .nested_second import NestedSecond
 from .pagination import Pagination
 from .parameters import Parameters
 from .requestbodies import RequestBodies
@@ -42,6 +45,9 @@ class SDK:
     r"""Endpoints for purely testing valid generation behavior."""
     globals: Globals
     r"""Endpoints for testing global parameters."""
+    nested: Nested
+    nested_first: NestedFirst
+    nested_second: NestedSecond
     pagination: Pagination
     r"""Endpoints for testing the pagination extension"""
     parameters: Parameters
@@ -149,6 +155,9 @@ class SDK:
         self.flattening = Flattening(self.sdk_configuration)
         self.generation = Generation(self.sdk_configuration)
         self.globals = Globals(self.sdk_configuration)
+        self.nested = Nested(self.sdk_configuration)
+        self.nested_first = NestedFirst(self.sdk_configuration)
+        self.nested_second = NestedSecond(self.sdk_configuration)
         self.pagination = Pagination(self.sdk_configuration)
         self.parameters = Parameters(self.sdk_configuration)
         self.request_bodies = RequestBodies(self.sdk_configuration)