+++ typescript-client-sdk/src/sdk/sdk.ts
@@ -13,6 +13,9 @@ import { Globals } from "./globals";
 import * as errors from "./models/errors";
 import * as operations from "./models/operations";
 import * as shared from "./models/shared";
+import { Nested } from "./nested";
+import { NestedFirst } from "./nestedfirst";
+import { NestedSecond } from "./nestedsecond";
 import { Pagination } from "./pagination";
 import { ParametersT } from "./parameters";
 import { RequestBodies } from "./requestbodies";
@@ -123,8 +126,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "0.1.0";
-    sdkVersion = "1.31.0";
-    genVersion = "2.91.2";
+    sdkVersion = "1.31.1";
+    genVersion = "2.91.4";
     globals: any;
 
     public constructor(init?: Partial<SDKConfiguration>) {
@@ -167,6 +170,9 @@ export class SDK {
      * Endpoints for testing global parameters.
      */
     public globals: Globals;
+    public nested: Nested;
+    public nestedFirst: NestedFirst;
+    public nestedSecond: NestedSecond;
     /**
      * Endpoints for testing the pagination extension
      */
@@ -266,6 +272,9 @@ export class SDK {
         this.flattening = new Flattening(this.sdkConfiguration);
         this.generation = new Generation(this.sdkConfiguration);
         this.globals = new Globals(this.sdkConfiguration);
+        this.nested = new Nested(this.sdkConfiguration);
+        this.nestedFirst = new NestedFirst(this.sdkConfiguration);
+        this.nestedSecond = new NestedSecond(this.sdkConfiguration);
         this.pagination = new Pagination(this.sdkConfiguration);
         this.parameters = new ParametersT(this.sdkConfiguration);
         this.requestBodies = new RequestBodies(this.sdkConfiguration);